Use byte arrays instead if net.IP
net.IP is not hashable, when using net.IP, when we choose to hash it
it needs a conversion to a string which is a allocation. Instead,
convert it to a byte array which requires zero allocs when converting
from net.IP which is a typedef for a byte slice. The conversion from
byte slice to a byte array is a tricking copy into thinking that the
byte array is actually a slice.