Don't create too many timers
Try and limit the use of timers in collector as well as reporter.
The previous implementation in the collector was creating one timer per
connection (packet) observed which in turn required a goroutine to
handle the timer firing. This was causing a slow heap increase to the
bottom line. Now, since we iterate through all the entries anyway when
we have to report them we check and expire entries then.