Do not mount cnx-apiserver cert/key paths
When not using custom generated cnx-apiserver cert/keys for the purposes
of communication between CNX-APIServer and K8s-APIServer with verifiable TLS
certificates, do not mount the /code/apiserver.local.config/certificates
path.
Its the default path used by the cnx-apiserver to generate and write the
apiserver cert/key pair when none present. Starting 1.9.4 (with the
security patch) it seems a Secret Volume when mounted comes strictly in
readOnly mode.