Generalize and simplify code for log group and stream naming
The code as it stood was a bit confusing, because the general Felix
config struct seemed to indicate that the defaults for LogGroupName and
LogStreamName were "", and then the actual defaults were specified by
constants elsewhere.

Also, if a user chose to configure CloudWatchLogsLogGroupName or
CloudWatchLogsLogStreamName, they couldn't still get the useful
<cluster-guid> and <felix-hostname> substitutions - and we can easily
make the code more uniform in a way such that they _can_ still get those
substitutions.