Fix commenting of CloudWatchLogs settings
Defaults for LogGroupName and LogStreamName were different from what
Felix actually does.  (We could change Felix, instead of this change,
but as we've been working with this for a while now, I guess it's likely
that people have seen and decided that they're happy with the names that
Felix is now generating.)

Felix now checks and sets (if necessary) RetentionDays each time it
starts up, so the comment about LogGroup creation time was wrong.