Pin to scapy 2.4.0, to avoid flake with TXT record
I hit a flake with the test that sends a TXT record, like what is
described at https://github.com/secdev/scapy/issues/1895.  That issue
indicates that 2.4.0 was better, so let's pin back to that.  In any case
it's better for us to be pinned than floating.

Also we don't need ipython for the usage in these FV tests, so skip
installing that.