Fix BGP peering and namespace deletion hang in test_rr
The end of test_rr changed the BGP setup - in a way that meant that the
cluster would soon lose general connectivity - and then initiated
cleanup by deleting the namespace.  There was then a race between the
namespace cleanup and the loss of connectivity, in particular
connectivity between the kube-controller-manager and the CNX API server.

It wasn't intentional to change the BGP setup in that broken way, so the
fix here is to fix that, with all of the other nodes peering with the
RR.