Don't initialize ancestorNames on empty nameData
When creating an empty nameData, to represent an RHS name that we expect
to get values for shortly, we don't need to initialize ancestorNames
with the parent name, because that will happen anyway when GetDomainIPs
is called for the parent name or any of the parent name's
ancestors.  (If GetDomainIPs isn't called for any of those, the empty
nameData isn't of interest yet to the IPSets Manager.)

It's a useful simplification _not_ to initialize ancestorNames here,
because then the updating of ancestorNames fields is entirely within
GetDomainIPs, which is easier to understand.