Fix branched DNS graph bug
The key point is that GetDomainIPs now _adds_ to the namesToNotify field
of relevant names, instead of just _replacing_ it.  To facilitate that,
namesToNotify is now a set.Set (instead of a slice).

At the same time, the code inside GetDomainIPs is simplified, because
there's no need to add any intermediate names to namesToNotify.
(I.e. intermediate names in the chain for which there has never been a
GetDomainIPs() call.)

I think this leaves us with two points needing further thought and work:

- When should we _remove_ names that are no longer of interest from
  namesToNotify fields?

- When can we delete entries in the s.mappings map?  (In particular,
  I suspect the `len(nameData.values)+nameData.namesToNotify.Len() == 0`
  condition in processMappingExpiry is wrong.)

It's not functionally a problem to leave outdated information in place,
but it means we're using more memory than needed.