Don't do unnecessary fsync when saving DNS mappings file
On my slow-ish laptop, saveMappingsV1() was taking up to 5 seconds each
time, while holding the domain info store lock.  That in turn was
causing surprisingly high latency between the sending of a DNS response,
and Felix's conversion of that to an ipset update, and hence FV test
failures.

I think this was caused by the `f.Sync()` call - which actually syncs
the _entire_ filesystem, not just the particular file - and that that
kind of sync isn't necessary for our use case.  Our use case is to
preserve DNS mappings on a given compute host when Felix restarts, and
fsync - which is only relevant when the operating system as a whole
crashes - makes no difference in that case.  If a compute host's whole
operating system does crash, it's acceptable that DNS mappings will have
to be relearned after the host and Tigera are restarted.

Therefore it's better to remove the `f.Sync()`.  It's possible that
leaving it in would cause obscure timing-dependent problems elsewhere in
Felix.