TLS tunnel is associated with a cluster
After a cluster is registered in voltron, we can retrieve generated creds
and use them to open a tunnel to voltron, which is then associated with
the registered cluster.

The private keys are never keept unless WithKeepClusterKeys option is
given, which is exclusively for testing and debugging.

fixes SAAS-97
fixes SAAS-52