Reduce aggressiveness of graceful shutdown packet loss tests
For some reason, on Semaphore, we see more packet loss on the
Debian base image that on Alpine, even with the same version
of strongswan.  Backing off the intensity of the test seems
to reduce the percentage loss, hinting at a resource
contention of some kind (maybe a libc difference?).