Avoid incorrect warning logs about CNAME loops
As reported in https://tigera.atlassian.net/browse/CNX-8724, with
ongoing DNS stress Felix was emitting lots of warning logs like this:

2019-07-23 14:08:32.855 [WARNING][46] domain_info_store.go 469: erick.com has a CNAME loop back to itself
2019-07-23 14:08:32.855 [WARNING][46] domain_info_store.go 469: hyenic.com has a CNAME loop back to itself
2019-07-23 14:08:32.855 [WARNING][46] domain_info_store.go 469: HDRedirect-LB5-1afb6e2973825a56.elb.us-east-1.amazonaws.com has a CNAME loop back to itself
2019-07-23 14:08:32.855 [WARNING][46] domain_info_store.go 469: honeydew.com has a CNAME loop back to itself
2019-07-23 14:08:32.855 [WARNING][46] domain_info_store.go 469: buyterbooks.palbin.com has a CNAME loop back to itself
...

Explanation is that xyz.com is often an alias for www.xyz.com, or vice
versa, so the `s.mappings` map may include:

- "xyz.com" -> CNAME: "www.xyz.com"
- "www.xyz.com" -> IP: 1.2.3.4

Then when we're iterating through `s.mappings` to match a wildcard like
`*.com`, we process "xyz.com" first, then "www.xyz.com"; or vice versa.
But the processing for "xyz.com" adds both "xyz.com" and "www.xyz.com"
to collectedNames, and then the processing for "www.xyz.com" emits a
warning log because "www.xyz.com" is already in collectedNames.

Fix is for each iteration of the wildcard loop to use a fresh set of
collectedNames.