Match SplitProfileRevision() to SplitNetworkPolicyRevision().
After adding Profile to AAPI server, we noticed a constant flow
of 'ResourceVersion is not valid: 0' every second in logs. Based on
the explanation here --
https://github.com/kubernetes/client-go/issues/366#issuecomment-364627356 --
it appears that k8s has a special meaning for ResourceVersion value 0
compared to "". It can be seen here --
https://github.com/kubernetes/kubernetes/blob/f6bc0ea03e5a9c392893f45687cc131eb8daae3f/staging/src/k8s.io/client-go/tools/cache/reflector.go#L166

Handling for a possible string with no "/" in SplitProfileRevision()
appears to have fixed the issue.