accept only tunnels from previously registered clusters
List previously registered clusters before setting up a watch to get
lasr version of the list so that the watch can start from that point.

As a side-effect, it provides the list of previously registered clusters
when voltron restarts, so we only accept tunnels from such clusters,
that is, if a cluster was previously deleted, it is not possible to use
its certs to connect a tunnel.

We skip matching the cert until SAAS-222

fixes SAAS-221
fixes SAAS-235