redo cluster locking in voltron
The RLock was held during the whole time when a request is proxied. That
is not necessary and can block other accesses to the cluster for long
periods of time if the request takes a long time, e.g. when the requests
is a k8s watch. For instance, it is not possible to remove a cluster
during this period.

We only need to grab the RLock to get access to the tunnel or proxy. They
are both thread safe, therefore we do not need to hold the lock when we
use them. Removing or updating the cluster would stop the live
connections and all users of tunnel, the proxy and the connections will
exit safely.

fixes SAAS-226