[SAAS-175] Adding basic authentication (#160)
* [SAAS-175] Adding basic authentication

K8s Api does not provide a method to authenticate an user and password. In order to validate the user and password, an anonymous config will be created and populated with the provided user name and password. This config will be used to create a K8s client API that performs a SelfSubjectAccessReview for a feke resource.

* [CODE REVIEW] Move k8sConfig as an Authentication option

* [Code review] Rename interfaces

* [CODE REVIEW] Removing config as a parameter for server