K8ST: Work around unavailability of Docker repo
The kubeadm-dind-cluster image that we use in K8ST has a Docker repo
in its APT sources, and it appears that that Docker repo is now
unavailable, which causes an "apt update" to fail and scupper the
whole test.

We don't actually need any updates from that repo, so add
--fix-missing to work around that.  Also change apt to apt-get, as
that's best practice for scripting.