Open up network policies for Elasticsearch and Kibana
The operator needs access to Elasticsearch adn Kibana, however, since the operator is on the hostnetwork it's hard to create specific network policies for it. Since we're under a time crunch for 2.6.0 we need to just open up the policies instead implementing the changes required to allow for fine grained operator network policies.