Copy the elasticsearch package from the Operator project for creating users / roles
There are no changes to this package from what was approved in the Operator project. This is needed because we are moving the elasticsearch user / role creation from the Operator project to here, so that users / roles can be created in the managed clusters.

Note there are no tests for this, because there were no test before in the Operator project. If I add new code to this package then I will add tests for it then.