Update elasticsearch resource watches for elasticsearch / managed cluster controllers
When the api controller starts none of the elasticsearch resources that are needed to reconcile are available. Unfortunately, we also don't have the appropriate triggers for the watches, i.e. when es resources are added, when the es user secret becomes available.

This commit updates / adds watches for the elasticsearch resources so that the reconcile loops are triggered when the needed elasticsearch resources are available