Don't prepend .* to indices that don't have the date suffix
Right now the UI requests tigera_secure_ee_events*/search, which we translate to tigera_secure_ee_events.<cluster>.* as the name is usually <index-pattern>.<cluster>.<date>, however, the tigera_secure_ee_events index doesn't have the date suffix so the .* at the end of tigera_secure_ee_events.cluster.* matches a minimum of tigera_secure_ee_events.cluster. (notice the period), which doesn't exist

This PR adds a check to see if the index pattern matches known "dateless" indices

Note that we could just update this in the Managed to change the requests to send tigera_secure_ee_events instead of tigera_secure_ee_events*, but the idea behind any changes to es proxy were that the Manager should still work as is. It's likely that we will also change the API structure to remove the /_search endpoint, so we want to restrict the amount of work dedicated to tweaking the UI to match changes in ES proxy