(CNX-11541) Set VOLTRON_COMPLIANCE_INSECURE_TLS to true
The compliance certificate is generated within the compliance server pod itself, so voltron can't know what it is.

Note: for operator installs, the tigera operator generates the certificate and stores it in the secret, but this doesn't happen for helm installs. Since helm installs are going away (or at least going to wrap the operator install) we may as well keep the insecure tls behaviour that was previously used.