Fix IPv6 service advertisement ECMP route testing
We correctly get an ECMP route with link-local next hop addresses,
where the test code was expecting global next hop addresses.  I've
loosened the test so that it only checks the _number_ of next hops and
not the exact addresses.

See the comment in the test code for more detail.