Fix flaky test
One of the tests was trying to eventually conntect to the server, but it was using the same request body each time. The problem with this is that the http client close the request body when it's done, so only the first request was valid, and if that failed all subsequent requests would fail.