Improved hypothesis for DNS policy flake fix
Following further discussion with Shaun (@fasaxc), we think the
hypothesis at
https://github.com/tigera/felix-private/pull/678/commits/f1a4df3c153ef2aaf65f66615c1b1ec1c70c2e44
was wrong, because the logs from a failing test showed 3 wgets, of
which #1 and #3 failed, and #2 succeeded:

```
2020-04-29 17:51:09.632 [INFO][18054] utils.go 128: Creating Command. ... "wget", "-U", "firefox", "-T", "4", "microsoft.com"}
2020-04-29 17:51:13.784 [INFO][18054] dns_test.go 76: wget said:
wget: unable to resolve host address 'microsoft.com'
2020-04-29 17:51:14.784 [INFO][18054] utils.go 128: Creating Command. ... "wget", "-U", "firefox", "-T", "4", "microsoft.com"}
2020-04-29 17:51:16.223 [INFO][18054] dns_test.go 76: wget said:
[successful]
2020-04-29 17:51:16.223 [INFO][18054] utils.go 128: Creating Command. ... "wget", "-U", "firefox", "-T", "4", "microsoft.com"}
2020-04-29 17:51:20.366 [INFO][18054] dns_test.go 76: wget said:
wget: unable to resolve host address 'microsoft.com'
```

And all of these were after all the workload and DNS policy
configurations.  A race with installing that policy cannot explain why
wget #3 failed after wget #2 had succeeded.

Apparently there is a known "parallel DNS" bug.  Some DNS client
implementations send their requests in parallel from the same source
port.  When you combine that with Linux's NAT implementation, it can
cause collisions in the conntrack table and some packets get lost.

We definitely have a NAT (or MASQUERADE) for this test; it's needed to
contact DNS servers and microsoft.com:

```
		// Allow workloads to connect out to the Internet.
		felix.Exec(
			"iptables", "-w", "-t", "nat",
			"-A", "POSTROUTING",
			"-o", "eth0",
			"-j", "MASQUERADE",
		)
```

So, if this bug is hitting us, and causing some UDP DNS packets to be
lost:

- it makes sense that `--dns-timeout=1` (added in yesterday's PR for a
  different reason) would help

- we should also add `--random-fully` to the MASQUERADE rule; this
  reduces the SNAT collision problem considerably.

Finally, the `--no-dns-cache`, that we also added yesterday, really
makes no sense, given the new hypothesis, so we revert that here.