Don't copy label map in wep conversion
Previously the label map was shared between the pod and the wep in PodToWorkloadEndpoint, but this means updates to the labels in one component are update in the other component. This really causes issues when using multi interface mode multus and the pod already has a label (thus the map is shared among all the weps and the labels will be the same for all WEPs for the pod).