[SAAS-1066] Felix starts/stops a packet capture (#748)
Once a PacketCapture is created via the API, Felix will match the workload endpoints in the calculation graph and send an PacketCaptreUpdate the the internal data plane. The internal data plan will start a capture once the matching workload endpoints and interfaces are seen as active. All live traffic will be stored on disk as pcap files under /var/log/calico/pcap. The capture files will be rotated using size, time and rotated files prune to a max number of files. Rotation can be configured via FelixConfig.
A packet capture is stopped via a delete api call. This will trigger a PacketCaptureRemoval to be sent to the internal data plane tostop capture on any active workload endpoint.