Fix DNS UT flake
This flake is a rare race condition, I believe.  It occurs when
`checkMonitor` is expecting a domain change signal for a domain
name (e.g. "*.microsoft.com") and then that `GetDomainIPs` for that
domain name will return a particular set of IPs, and what actually
happens is that:

1. The domain name is signaled twice (or more times) very quickly.

2. After the first of those signalings, `GetDomainIPs` does not return
the expected set of IPs.

3. But after the last of those signalings, `GetDomainIPs` _does_
return the expected set of IPs.

The fix is to move the IPs check inside the `Eventually` in
`checkMonitor`.

For the record, the symptom was like this:

• Failure [0.012 seconds]
Domain Info Store with monitor thread [It] microsoft case
/go/src/github.com/projectcalico/felix/dataplane/linux/domain_info_store_test.go:274
  Expected
      <[]string | len:0, cap:0>: nil
  to equal
      <[]string | len:1, cap:1>: ["104.75.174.50"]
  /go/src/github.com/projectcalico/felix/dataplane/linux/domain_info_store_test.go:257