Golang 1.15 upgrade / generate certificates for tests
This commit fixes certificate issues that happened with upgrading to golang 1.15 where the tests fail if the common name is used instead of the SANs

I've tried to move away from using certificate files in our tests, and instead, generate the certificate objects we need for the tests. This allows a user to clearly see how a certificate was created for a test and makes it very easy to tweak a certificate for a test, while reducing the number of different certificates we need to store for tests.

The tests in server/http_traffic_server_test are actually tests for the certificates, making sure different formats are valid, so I left them as is.

Some tests I decided to leave, like the ones in pkg/tls, because I've spent too much time on this already at the moment.

I've cleaned up the tests in server_test because quite a few of them weren't actually testing anything and are using the WatchK8sWithSync, which has a race condition in it and in general is not a good thing to use because it is test code that has been inserted into production code. I've also started moving away from the test/k8s fake client as it's not needed and over developed, we just need the fake.NewSimpleClientset().ProjectcalicoV3().