Use Epoch seconds for Elasticsearch queries when given absolute times
Currently sending a RFC3339 time stamp does not work for the flowLogs endpoint as the format is not included in the Elasticsearch query, so Elasticsearch assumes the timestamp is in epoch seconds. This PR just converts the timestamp that is passed to Elasticsearch into epoch seconds.

This is not broken in production because the Manager uses an Elasticsearch format relative to the current time, i.e. now-X, and that code path is not broken.