Refactor PolicyHit to fix not identifying staged policies
This commit does the following:
- Parse policy name a little deeper (breaking down the policy name into further parts) and fix not identifying staged policies
- Make PolicyHit an immutable object as they are representations of objects that aren't changeable.
- Remove using ActionFlags within a PolicyHit, this is something needed specifically for PIP and isn't necessarily useful in representing a legitimate policy hit

For the first point the format for staged policy logs differs between kubernetes network policies and calico network policies, where kubernetes network policy names start with the "staged" prefix, e.g. staged:knp.default.policyName, and calico network policy names have them after the tier prefix, e.g. tierName.staged:policyName.

For the second point, PolicyHits are representations of logs that have already happened, and are unchangeable in general, so it seems fitting to make a parsed policy log immutable and code that needs to modify or aggregate a log can create a new PolicyHit. Convenience "Set" functions have been added to PolicyHit to set values on a copy of the underlying policy and return that copy, keeping the original in tact to maintain the immutable properties of a PolicyHit.