Use refactored immutable lma PolicyHit
This commit updates the PolicyHit implementation used to the new one in lma. The biggest change is that PIP know log creates PolicyHits that have multiple actions, that are then rendered into multiple policy logs, but instead creates a single policy for each action, and each PolicyHit is guaranteed to generate a single policy log line.