Return bool not status for authorize
Returning a status alongside an error (which may be returned even if there wasn't an error) makes the use of this function confusing and difficult. For instance, a 403 (unauthorized) is not an error in regards to this function, it's an expected response, and the use of 0 as a return status breaks http return status assumptions. Simplifying this function to simply return true or false (authorized or unauthorized) and an error if something was wrong makes this function simpler to use.