Add CanListEndpoint() function to FlowHelper
CanListEndpoint takes an endpoint type and a namespaces and figures out what exactly it needs to authorize list capabilities for based on that. This was generalized as there's a couple of places that need this general logic and it shouldn't be duplicated as changes in one place would need to be cascaded to the duplicates.