collector: plugin interface tweaks
Methods that return the info channes are named like the info they return
to allow implementing all plugins by a single object - combining the
interfaces.

Start() returns an error to be able to fail starting the collector if
any of the plugins fails to start. This allows the user of collector to
report and quit.