bpf: generate policy verdict events
When a rule is inserted in the lookups cache, another 64bit ID is
generated for the rule. The extra ID is generated only in the BPF mode
when this feature is explicitly turned on.

BPF ep manager passes enough information to the policy program builder
to generate the same rule ID used by iptables/NFLOG. This id can be
used to lookup the additional 64bit id, which is then encoded in the
policy program. These 64bit ids are returned to userspace after a packet
is processed by the policy program and a verdict is reached.