BPF support for flowlogs
PacketInfo plugin for the collector uses the lookups cache to translate
the 64bit id to the rule ID. From this point on, the path is unified for
all dataplanes.

Conntrack has an extra scanner, which acts as a collector plugin and
generates ConntrackInfo.

Since BPF does not support HEP policy, no events are generated in such
cases and the test cannot expect the same result as in the iptables
case. The test are amended to expect private network in aggregation
instead.