Kibana OIDC login with native elasticsearch users
This commit adds the logic to log in OIDC users to Kibana. It does this by getting the OIDC users corresponding Elasticsearch user, looking up the password for that user stored in one of our secrets, making the login request to Kibana, setting the cookie Kibana would have set, then redirecting to Kibana.