Keep .d files so that we don't always do a complete rebuild
A while ago, I merged a couple of commits,
https://github.com/tigera/felix-private/commit/a500e213 and
https://github.com/tigera/felix-private/commit/bc26e37c, whose effect
is that we now get a complete rebuild of all the BPF C code for any
BPF-related `make` invocation; and the C code being rebuilt means that
various dependent containers are rebuilt too, even if nothing
contributing to them has actually changed.

This wastes a lot of time, in particular when iterating on UT or FV
work.

So why did I merge those commits?  The commit message for the first of
them says:

    Avoid missing bpf.h problem, often seen with 'make ut'
        make: *** No rule to make target '/usr/src/linux-headers-5.4.0-0.bpo.3-common/include/uapi/linux/bpf.h', needed by 'connect_balancer.d'.  Stop.
        Makefile:261: recipe for target 'build-bpf' failed
    According to Slack discussion the solution is to delete the .d files,
    so let's always do that as part of the 'build-bpf' target.

The scenario was this:

- You do a build, and that populates connect_balancer.d (for example)
  with the header file dependencies for connect_balancer.c

- The "system" dependencies, like
  /usr/src/linux-headers-5.4.0-0.bpo.3-common/include/uapi/linux/bpf.h,
  can have specific and/or versioned paths that depend on the version
  of calico/go-build that you are using.

- You do something, e.g. switch to a different branch, that causes the
  calico/go-build version to change, and some of the "system"
  dependencies in the *.d files now do not exist inside the
  calico/go-build container.

- Now, when you rebuild, you get the "No rule" error above, because
  connect_balancer.d says that it needs a specific version of bpf.h,
  but that version is not available.

However that is a relatively rare scenario, and I think is becoming
rarer, for two reasons.

- Our calico/go-build support for BPF building has matured and is
  likely to change less frequently in future.

- The system dependencies that we have now (go-build v 0.50) do not
  have specific kernel versions any more.

So, on balance, we should just revert those commits and not
automatically wipe the .d files.

If we find that we do start seeing the "No rule" error again, we
should address it in a better way by adding a Makefile pattern rule
that matches the system header paths, and that does the CC invocation
to regenerate the .d files.  I'm not doing that here because I haven't
found a current way - despite a few attempts - to repro the "No rule"
scenario.