bpf: if events.New() fails do not panic
Also do not create any other components that depend on the events.

The primary usecase is when FlowLogsCollectProcessInfo is enabled but
the kernel is too old to support it.

When we run in full BPF mode, not having some components like the
bpfEventPoller would likely lead to a failure/panic anyway. But that is
probably more right than wrong.