bpf: account bytes and packets corretly with tunnels
Since when the traffic is tunneled, BPF program on the host iface sees
it first and marks it as seen before another program sees the packet as
decaped. Only then we can account for the bytes and packets.
Unfortunately, we need to mark the packets as seen otherwise they would
get dropped as something we missed.

Needs to be done for tunnels that preserve the packet, like IPIP and
unlike wireguard.