Avoid/downgrade "Already have outstanding DNS request" WARNING logs
We've been seeing spam quantities of this log on a BPF-enabled steady
state Calico Enterprise rig:

    [WARNING][54] felix/domain_info_store.go 782: DNS-LATENCY: Already have outstanding DNS request with ID <..>

I believe there are two root causes, and in both cases we can safely
avoid or downgrade the warning without any impact on DNS log
correctness.

Cause 1 is that when a DNS request goes from a local pod (client) to
another local pod (e.g. coredns) on the same node, the BPF dataplane
reports the DNS request up to this (domain_info_store) code twice.
We do have a related tangible problem here, in that the corresponding
_response_ is also reported up twice, which means that it gets
DNS-logged twice, once with latency and once without.  We still need
to solve that, but in the meantime we don't lose anything by just
downgrading the warning log when seeing the same _request_ twice.

Cause 2 is that DNS IDs - which we use to correlate requests and
responses - are only 16-bit, so it's plausible that we will
occasionally see the same ID being used at the same time from
different clients, which will confuse our correlation code.  To fix
that, we should scope the DNS ID by the client IP.

Cause 2 is speculative, but I think the stats from an actual rig
support it.  Here are counts of the occurrences of the WARNING log
from a 5 node rig, from the 2 nodes with coredns pods:

    137095 calico-node-g5f2w coredns
      2482 calico-node-q8dvr coredns

and from the 3 other nodes:

        18 calico-node-f4nmq
         4 calico-node-hdhcx
        20 calico-node-v47s2

The coredns node counts indicate the scale of how many DNS exchanges
have occurred.  I think the non-coredns node counts are small enough,
in comparison, for cause 2 to be a plausible explanation for them.