Unify meanings and usages of CT state ifindex field
Before the BPF DNS policy work, calico_ct_leg.ifindex was always set
to the ingress interface index of packets on that leg.

- For a leg coming from a workload or outside the
host (CALI_F_TO_HOST), this is the same as the index of the interface
that the TC program is on.

- For a leg coming from the host (CALI_F_FROM_HOST), it's different.
Actually we haven't checked if Linux populates an ingress interface
index at all, in this case, but it definitely shouldn't be the
interface index of the TC program that the packet is _egressing_
through.

Prior to DNS work, we actually only needed calico_ct_leg.ifindex to be
set in CALI_F_TO_HOST cases.  It's used in order to do RPF on incoming
packets (or to let iptables do RPF), and for looking up the outgoing
interface index when a packet is being bypassed; and neither of those
applies to legs coming from the host.

Now, for DNS I want a way of determining "is the current TC program on
the interface where the CT state (for this packet) was first created?"
But I only need that when the CT state was created in a CALI_F_TO_HOST
context, so the condition for DNS purposes can be "CT state was
created in a CALI_F_FROM_HOST context, or on the same interface as the
current TC program."

Therefore the key point of this change is _not_ to set
calico_ct_leg.ifindex when creating a CT state in a CALI_F_FROM_HOST
context.  Instead it's left as CT_INVALID_IFINDEX (0).  Then all of
the pre-DNS cases still work, and the DNS condition can be

   (CT ifindex == CT_INVALID_IFINDEX) || (CT ifindex == this TC program)