Remove Elasticsearch configuration reconcile on start up, rely on add events
This commit removes the Reconcile at the beginning of the elasticsearch config controllers run, and instead relies on the initial reconcile it does when the watch is first created. I've made sure that the reconciler watches for add events, which means that when the controller starts it will try to reconcile all the resources it's watching, triggering the intial reconcile in a more natural way.

Doing this allows the reconcile to fall back on the retry logic of the worker if there's a failure, unlike before where we would only attempt the initial reconcile once.

I've also allowed the max requeue attempts to be configurable and change that value to 20 for the elasticsearch configuration. The reason for this change is because the events that trigger this reconciliation are very frequent, so we don't want to lose the event on a temporary outage (like when elasticsearch restarts).