Use env variables instead of k8s to created Elasticsearch client
This replaces dynamically grabbing elasticsearch credentials / certificates from k8s to create the client and instead gets them from the environment. This is the standard way of populating the credentials, and allows us to mount secrets with these values.

I've added a elasticsearch ClientBuilder, so that we can populate it with the required configuration then later create a client when elasticsearch is up.