Create and use a more reliable Netcat-like client
Every now and then a test fails because of the nc client dying
prematurely with exit code 1.  Adding "-v" to the invocation
reveals (in all of the cases that I've seen after adding "-v"):

    nc: too many output retries

So I went to the code to try to understand this.

It turns out, Netcat does not select on the TCP socket to find out
when it can _write_ more data to it.  Instead it just retries writing
in a tight loop, and if that fails around 8000 times, gives up with
the error message as above.

So in a dual ToR resilience test, where it is likely the TCP socket
will block for a period of time, we are pretty likely to see Netcat
giving up in this way.

Therefore here we write and use a simple Netcat client replacement
using golang's io.Copy(), which is a wrapper around sendfile() and I
presume has more sensible write-blocking behaviour.