Watch Management cluster secrets / config maps in managed cluster elasticsearch controllers
Previously I removed the managed cluster initialization reconciler so that we could rely on the natural flow of k8s secrets / config maps being rolled out, but if there are no secrets or config maps in the managed clusters nothing will trigger the elasticsearch credentials controller to reconcile. This commit makes sure the elasticsearc credentials controller watches the management cluster secrets, config  maps and elasticsearch itself so that the reconciliation for the managed clusters will be triggered when there's changes.

This also removes the need to restart the elasticsearch credentials controllers when things change in the management cluster, since each elasticsearch credentials controller is resyncing when the secrets in the management cluster change.