ConntrackInfoReader uses batches of ConntrackInfo
When conntrack table is large, it takes significant amount of time to
push individual ConntrackInfo through a channel, especially when
Felix's CPU time is limited and the system is overloaded. In case of BPF
that would slow down cleaning of the table, which increases the load as
many unexpired entries are visited multiple times.

Previously, iptables handled the entire content of conntrack in a single
go, this is pretty much return to that state.

Windows follow the same path, if the send into the ConntrackInfoChan is
blocked, the batch grows to a certain point. Internally, no pointers are
used as that may create unnecessary load on GC with many small items
allocated just to be deallocated soon after.

BPF splits the conntrack into batches of 256 and if it were to block,
the batch also grows, unbounded. This batching allows the collector to
interleave handling of other stuff with handling ConntrackInfos.