Fix simulation of BPF CT state for DNS scapy tests
Situation here was as follows.

DNS scapy tests work by
- Creating the CT state that would have been created by workload sending a DNS request (to a server on a different node).
- Sending in a matching DNS response.

With host IP present (which it is because of the current merge), the
DNS response passes through BPF TC programs for
- eth0 (ingress)
- the workload interface (ingress)

The TC program for eth0 creates a new CT state, and this loses some of
the information - notably ifindex on the opener leg - that we had on
the CT state that the test code created.  Loss of that information
means that the DNS response doesn't get reported when it passes
through the TC program for the workload interface.

The bug was test code not setting Whitelisted: true when creating the
CT state (to match the CT state that would have been created by the
DNS request passing through).  When the Whitelisted fields are
correctly set, a lot of processing is skipped in the eth0 program,
including the part that creates a new CT state.