Change egress gateway host-side addr from 169.254.1.1 to 169.254.1.2
We need to assign _some_ address to the host-side of the veth for an
egress gateway pod, so that RPF checking works for that pod.

Previously we assigned 169.254.1.1, but we also use 169.254.1.1 as the
default gateway address in each pod's routing table, and two of our
customers have now observed a conflict between those two uses of the
same address.

The conflict arises if `arp_ignore` is set to a value other than its
default; from
https://www.kernel.org/doc/Documentation/networking/ip-sysctl.txt:

    arp_ignore - INTEGER
    	Define different modes for sending replies in response to
    	received ARP requests that resolve local target IP addresses:
    	0 - (default): reply for any local target IP address, configured
    	on any interface
    	1 - reply only if the target IP address is local address
    	configured on the incoming interface
    	2 - reply only if the target IP address is local address
    	configured on the incoming interface and both with the
    	sender's IP address are part from same subnet on this interface
    	3 - do not reply for local addresses configured with scope host,
    	only resolutions for global and link addresses are replied
    	4-7 - reserved
    	8 - do not reply for all local addresses
    	The max value from conf/{all,interface}/arp_ignore is used
    	when ARP request is received on the {interface}

When a non-egress-gateway pod wants to send to anywhere, its routing
table says to go via 169.254.1.1, so it then ARPs for 169.254.1.1, and
we rely on proxy_arp (which we configure ourselves) so that the Linux
host responds with "yes, 169.254.1.1 is here".  But it sounds like if
arp_ignore is set to 1, and 169.254.1.1 is actually configured on a
different host interface - as it is when there is an egress gateway
pod - that prevents the proxy_arp response that we rely on.

One scenario that leads to arp_ignore=1 is kube-proxy's strict-arp
setting; from
https://kubernetes.io/docs/reference/command-line-tools-reference/kube-proxy/:

    --ipvs-strict-arp
    Enable strict ARP by setting arp_ignore to 1 and arp_announce to 2

In turn, the strict-arp setting is required for MetalLB.

There may be other scenarios that lead to 169.254.1.1 conflicts, but
this one is enough justification on its own for a simple change to use
a different address.